+++ chirho/dynamical/handlers/__init__.py
@@ -1,6 +1,6 @@
 from ..internals.backend import Solver  # noqa: F401
 from .dynamical import SimulatorEventLoop  # noqa: F401
-from .interruption.array_observation import StaticBatchObservation  # noqa: F401
+from .interruption.interruption import StaticBatchObservation  # noqa: F401
 from .interruption.interruption import (  # noqa: F401
     DynamicInterruption,
     DynamicIntervention,
