--- chirho/dynamical/handlers/interruption/__init__.py
@@ -1,2 +1,0 @@
-from . import array_observation  # noqa: F401
-from . import interruption  # noqa: F401
